Added pkclick/pgsfile.py and moved vcffile to its own file
